requirements.txt
streamlit
pandas
altair
tabulate
python-dotenv
openai
langchain
langchain-openai
langchain-experimental